requirements.txt
Django==1.11
django-filter==2.1.0
django-markdown-deux==1.0.5
django-markdownx==2.0.28
django-suit==0.2.26
djangorestframework==3.9.2
gunicorn==19.9.0
Markdown==3.0.1
markdown2==2.3.7
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
